feat: add hook for aggregated store data
