+++ src/services/index.ts
@@ -8,3 +8,4 @@ export * from './useProductPricing';
 export * from './useCart';
 export * from './useOrders';
 export * from './useUser';
+export * from './useStoreData';
